requirements.txt
streamlit
groq
python-dotenv
pypdf
streamlit-mic-recorder

